refactor: move logos constants in separate file
