+++ src/components/BrandCarousel/BrandCarousel.tsx
@@ -4,7 +4,7 @@ import Image from 'next/image';
 import { Carousel, CarouselContent, CarouselItem } from '@/components/ui/carousel';
 import Autoplay from 'embla-carousel-autoplay';
 
-import { brandLogos } from '@/constants/brands';
+import { brandLogos } from '@/constants/constants';
 
 const BrandCarousel = (): JSX.Element => {
   return (
