+++ src/constants/constants.ts
@@ -7,3 +7,11 @@ export const brandLogos = [
   'nike.png',
   'puma.png'
 ];
+
+export const socialLogos = [
+  'youtube.png',
+  'facebook.png',
+  'instagram.png',
+  'pinterest.png',
+  'twitter.png'
+];
