--- src/constants/socials.ts
@@ -1,7 +1,0 @@
-export const socialLogos = [
-  'youtube.png',
-  'facebook.png',
-  'instagram.png',
-  'pinterest.png',
-  'twitter.png'
-];
